requirements.txt
colorama==0.4.6
python-dotenv==1.2.1
tqdm==4.67.1
